airflow-gcp/requirements.txt
pyarrow
dbt-core
dbt-bigquery
apache-airflow-providers-google
